Fix crash 'double free'
The "delete" behovior in C++ will destroy the pointer(first destroy),
and the destruct function of class will call the "wlr_xxxx_destroy"
function(second destroy) to destroy the pointer. So, we can't using
C++ style API to destroy the wlr_xxxx pointer.

Add the "destroy" function to class and mark the destruct function
of class to private.

Allow the "dev" is nullptr in QWCursor::move.